requirements.txt
pandas
numpy
sentence-transformers
scikit-learn
python-dotenv
rouge-score
sacrebleu